+++ src/collective/cover/behaviors/interfaces.py
@@ -1,10 +1,11 @@
 # coding: utf-8
 from collective.cover import _
-from plone.directives import form
+from plone.autoform.interfaces import IFormFieldProvider
 from plone.supermodel import model
 from zope import schema
 from zope.interface import alsoProvides
 from zope.interface import Invalid
+from z3c.form import validator
 
 
 class IRefresh(model.Schema):
@@ -25,7 +26,18 @@ class IRefresh(model.Schema):
         default=300,
     )
 
-alsoProvides(IRefresh, form.IFormFieldProvider)
+class TimeToLiveValidator(validator.SimpleFieldValidator):
+
+    """z3c.form validator class for time to live"""
+
+    def validate(self, value):
+        """Validate the value on time to live input"""
+        if value <= 0:
+            raise Invalid(_(u'Value must be greater than zero.'))
+
+validator.WidgetValidatorDiscriminators(TimeToLiveValidator, field=IRefresh['ttl'])
+
+alsoProvides(IRefresh, IFormFieldProvider)
 
 
 @form.validator(field=IRefresh['ttl'])